fix virtual env and pre-commit

+++ minitorch/scalar_functions.py
@@ -105,34 +105,35 @@ class Mul(ScalarFunction):
     def forward(ctx: Context, a: float, b: float) -> float:
         # TODO: Implement for Task 1.2.
         """
-            Return the product of a and b.
+        Return the product of a and b.
 
-            Args:
-                a: float
-                b: float
+        Args:
+            a: float
+            b: float
 
-            Returns:
-                The product of a and b as a float.
+        Returns:
+            The product of a and b as a float.
         """
-        ctx.save_for_backward(a,b)
-        return operators.mul(a,b)
+        ctx.save_for_backward(a, b)
+        return operators.mul(a, b)
 
     @staticmethod
     def backward(ctx: Context, d_output: float) -> Tuple[float, float]:
         # TODO: Implement for Task 1.4.
         """
-            Return the derivatives of a and b with respect to the output.
+        Return the derivatives of a and b with respect to the output.
 
-            Args:
-                d_output: float
+        Args:
+            d_output: float
 
-            Returns:
-                The derivatives of a and b with respect to the output as a tuple.
+        Returns:
+            The derivatives of a and b with respect to the output as a tuple.
         """
         (a, b) = ctx.saved_values
         dx_a = d_output * b
         dx_b = d_output * a
         return (dx_a, dx_b)
+
 
 class Inv(ScalarFunction):
     "Inverse function"
@@ -141,13 +142,13 @@ class Inv(ScalarFunction):
     def forward(ctx: Context, a: float) -> float:
         # TODO: Implement for Task 1.2.
         """
-            Return the inverse of a.
+        Return the inverse of a.
 
-            Args:
-                a: float
+        Args:
+            a: float
 
-            Returns:
-                The inverse of a as a float.
+        Returns:
+            The inverse of a as a float.
         """
         ctx.save_for_backward(a)
         return operators.inv(a)
@@ -156,13 +157,13 @@ class Inv(ScalarFunction):
     def backward(ctx: Context, d_output: float) -> float:
         # TODO: Implement for Task 1.4.
         """
-            Return the derivative of a with respect to the output.
+        Return the derivative of a with respect to the output.
 
-            Args:
-                d_output: float
+        Args:
+            d_output: float
 
-            Returns:
-                The derivative of a with respect to the output as a float.
+        Returns:
+            The derivative of a with respect to the output as a float.
         """
         (a,) = ctx.saved_values
         return operators.inv_back(a, d_output)
@@ -175,13 +176,13 @@ class Neg(ScalarFunction):
     def forward(ctx: Context, a: float) -> float:
         # TODO: Implement for Task 1.2.
         """
-            Return the negation of a.
+        Return the negation of a.
 
-            Args:
-                a: float
+        Args:
+            a: float
 
-            Returns:
-                The negation of a as a float.
+        Returns:
+            The negation of a as a float.
         """
         ctx.save_for_backward(a)
         return operators.neg(a)
@@ -190,13 +191,13 @@ class Neg(ScalarFunction):
     def backward(ctx: Context, d_output: float) -> float:
         # TODO: Implement for Task 1.4.
         """
-            Return the derivative of a with respect to the output.
+        Return the derivative of a with respect to the output.
 
-            Args:
-                d_output: float
+        Args:
+            d_output: float
 
-            Returns:
-                The derivative of a with respect to the output as a float.
+        Returns:
+            The derivative of a with respect to the output as a float.
         """
         (a,) = ctx.saved_values
         return -1.0 * d_output
@@ -209,13 +210,13 @@ class Sigmoid(ScalarFunction):
     def forward(ctx: Context, a: float) -> float:
         # TODO: Implement for Task 1.2.
         """
-            Return the sigmoid of a.
+        Return the sigmoid of a.
 
-            Args:
-                a: float
+        Args:
+            a: float
 
-            Returns:
-                The sigmoid of a as a float.
+        Returns:
+            The sigmoid of a as a float.
         """
         sig = operators.sigmoid(a)
         ctx.save_for_backward(sig)
@@ -225,17 +226,18 @@ class Sigmoid(ScalarFunction):
     def backward(ctx: Context, d_output: float) -> float:
         # TODO: Implement for Task 1.4.
         """
-            Return the derivative of a with respect to the output.
+        Return the derivative of a with respect to the output.
 
-            Args:
-                d_output: float
+        Args:
+            d_output: float
 
-            Returns:
-                The derivative of a with respect to the output as a float.
+        Returns:
+            The derivative of a with respect to the output as a float.
         """
         (sig,) = ctx.saved_values
         d_sig = sig * (1 - sig)
-        return d_output * d_sig
+        return float(d_output * d_sig)
+
 
 class ReLU(ScalarFunction):
     "ReLU function"
@@ -244,13 +246,13 @@ class ReLU(ScalarFunction):
     def forward(ctx: Context, a: float) -> float:
         # TODO: Implement for Task 1.2.
         """
-            Return the ReLU of a.
+        Return the ReLU of a.
 
-            Args:
-                a: float
+        Args:
+            a: float
 
-            Returns:
-                The ReLU of a as a float.
+        Returns:
+            The ReLU of a as a float.
         """
         ctx.save_for_backward(a)
         return operators.relu(a)
@@ -259,13 +261,13 @@ class ReLU(ScalarFunction):
     def backward(ctx: Context, d_output: float) -> float:
         # TODO: Implement for Task 1.4.
         """
-            Return the derivative of a with respect to the output.
+        Return the derivative of a with respect to the output.
 
-            Args:
-                d_output: float
+        Args:
+            d_output: float
 
-            Returns:
-                The derivative of a with respect to the output as a float.
+        Returns:
+            The derivative of a with respect to the output as a float.
         """
         (a,) = ctx.saved_values
         return operators.relu_back(a, d_output)
@@ -278,33 +280,32 @@ class Exp(ScalarFunction):
     def forward(ctx: Context, a: float) -> float:
         # TODO: Implement for Task 1.2.
         """
-            Return the exponential of a.
+        Return the exponential of a.
 
-            Args:
-                a: float
+        Args:
+            a: float
 
-            Returns:
-                The exponential of a as a float.
+        Returns:
+            The exponential of a as a float.
         """
         exp_a = operators.exp(a)
         ctx.save_for_backward(exp_a)
         return exp_a
 
-
     @staticmethod
     def backward(ctx: Context, d_output: float) -> float:
         # TODO: Implement for Task 1.4.
         """
-            Return the derivative of a with respect to the output.
+        Return the derivative of a with respect to the output.
 
-            Args:
-                d_output: float
+        Args:
+            d_output: float
 
-            Returns:
-                The derivative of a with respect to the output as a float.
+        Returns:
+            The derivative of a with respect to the output as a float.
         """
         (exp_a,) = ctx.saved_values
-        return d_output * exp_a
+        return float(d_output * exp_a)
 
 
 class LT(ScalarFunction):
@@ -314,30 +315,31 @@ class LT(ScalarFunction):
     def forward(ctx: Context, a: float, b: float) -> float:
         # TODO: Implement for Task 1.2.
         """
-            Return 1.0 if a is less than b, else return 0.0.
+        Return 1.0 if a is less than b, else return 0.0.
 
-            Args:
-                a: float
-                b: float
+        Args:
+            a: float
+            b: float
 
-            Returns:
-                The result of the less-than comparison as a float.
+        Returns:
+            The result of the less-than comparison as a float.
         """
-        return operators.lt(a,b)
+        return operators.lt(a, b)
 
     @staticmethod
     def backward(ctx: Context, d_output: float) -> Tuple[float, float]:
         # TODO: Implement for Task 1.4.
         """
-            Return the derivative of a with respect to the output.
+        Return the derivative of a with respect to the output.
 
-            Args:
-                d_output: float
+        Args:
+            d_output: float
 
-            Returns:
-                0 as a float because the derivative of a less-than comparison is the derivative of a constant.
+        Returns:
+            0 as a float because the derivative of a less-than comparison is the derivative of a constant.
         """
         return (0.0, 0.0)
+
 
 class EQ(ScalarFunction):
     "Equal function $f(x) =$ 1.0 if x is equal to y else 0.0"
@@ -346,27 +348,27 @@ class EQ(ScalarFunction):
     def forward(ctx: Context, a: float, b: float) -> float:
         # TODO: Implement for Task 1.2.
         """
-            Return 1.0 if a is equal to b, else return 0.0.
+        Return 1.0 if a is equal to b, else return 0.0.
 
-            Args:
-                a: float
-                b: float
+        Args:
+            a: float
+            b: float
 
-            Returns:
-                The result of the equality comparison as a float.
+        Returns:
+            The result of the equality comparison as a float.
         """
-        return operators.eq(a,b)
+        return operators.eq(a, b)
 
     @staticmethod
     def backward(ctx: Context, d_output: float) -> Tuple[float, float]:
         # TODO: Implement for Task 1.4.
         """
-            Return the derivative of a with respect to the output.
+        Return the derivative of a with respect to the output.
 
-            Args:
-                d_output: float
+        Args:
+            d_output: float
 
-            Returns:
-                0 as a float because the derivative of an equality comparison is the derivative of a constant.
+        Returns:
+            0 as a float because the derivative of an equality comparison is the derivative of a constant.
         """
         return (0.0, 0.0)
